F_T30603 move DocumentBundle to core
- adjut use statements
