+++ demosplan/DemosPlanCoreBundle/ResourceTypes/SingleDocumentResourceType.php
@@ -14,7 +14,7 @@ namespace demosplan\DemosPlanCoreBundle\ResourceTypes;
 
 use demosplan\DemosPlanCoreBundle\Entity\Document\SingleDocument;
 use demosplan\DemosPlanCoreBundle\Logic\ApiRequest\ResourceType\DplanResourceType;
-use demosplan\DemosPlanDocumentBundle\Logic\SingleDocumentService;
+use demosplan\DemosPlanCoreBundle\Logic\Document\SingleDocumentService;
 use EDT\PathBuilding\End;
 use EDT\Querying\Contracts\PathsBasedInterface;
